rename : 폴더명 카멜케이스 변경
