--- src/libs/utils/dummy.txt
@@ -1,1 +1,0 @@
-유틸 더미
