+++ src/pages/sign/PasswordFind.tsx
@@ -4,7 +4,7 @@ import { styled } from "styled-components";
 import Input from "../../components/common/input/Input";
 import { BUTTON_OPTIONS, SIGNIN_OPTIONS } from "../../libs/constants/options/options";
 import Button from "../../components/common/button/Button";
-import { PopupPasswordFind } from "../../components/modal/UseSignPage/PopupPasswordFind";
+import { PopupPasswordFind } from "../../components/modal/useSignPage/PopupPasswordFind";
 
 const PasswordFind = () => {
   const [phoneNumValue, setPhoneNumValue] = useState("");
